requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validación y serialización
pydantic==2.5.0
pydantic-settings==2.1.0

# Autenticación y seguridad
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# IA y procesamiento de texto
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
whisper-openai==20231117
speech-recognition==3.10.0

# Procesamiento de audio
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# HTTP y requests
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Cache y sesiones
redis==5.0.1
aioredis==2.0.1

# Utilidades
python-dotenv==1.0.0
typing-extensions==4.8.0
email-validator==2.1.0

# Logging y monitoreo
loguru==0.7.2

# Testing (desarrollo)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Formateo y linting (desarrollo)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentación
mkdocs==1.5.3
mkdocs-material==9.4.8